migrate to separate resource function in AWS
