+++ cartography/intel/aws/resources.py
@@ -14,6 +14,7 @@ from . import cognito
 from . import config
 from . import dynamodb
 from . import ecr
+from . import ecr_image_layers
 from . import ecs
 from . import efs
 from . import eks
@@ -88,6 +89,7 @@ RESOURCE_FUNCTIONS: Dict[str, Callable[..., None]] = {
     "ec2:volumes": sync_ebs_volumes,
     "ec2:snapshots": sync_ebs_snapshots,
     "ecr": ecr.sync,
+    "ecr:image-layers": ecr_image_layers.sync,
     "ecs": ecs.sync,
     "eks": eks.sync,
     "elasticache": elasticache.sync,
